+++ mex/backend/graph/connector.py
@@ -1,4 +1,5 @@
 import json
+from collections.abc import Sequence
 from string import Template
 from typing import Annotated, Any, Literal, cast
 
@@ -28,8 +29,11 @@ from mex.common.models import (
     MEX_PRIMARY_SOURCE_IDENTIFIER,
     MEX_PRIMARY_SOURCE_IDENTIFIER_IN_PRIMARY_SOURCE,
     MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
+    RULE_MODEL_CLASSES_BY_NAME,
     AnyExtractedModel,
     AnyRuleModel,
+    AnyRuleSetRequest,
+    AnyRuleSetResponse,
     BasePrimarySource,
     ExtractedPrimarySource,
 )
@@ -155,11 +159,49 @@ class GraphConnector(BaseConnector):
             logger.debug("\n%s", message)
         return result
 
+    def _fetch_extracted_or_rule_items(
+        self,
+        query_string: str | None,
+        stable_target_id: str | None,
+        entity_type: Sequence[str],
+        skip: int,
+        limit: int,
+    ) -> Result:
+        """Query the graph for extracted or rule items.
+
+        Args:
+            query_string: Optional full text search query term
+            stable_target_id: Optional stable target ID filter
+            entity_type: List of allowed entity types
+            skip: How many items to skip for pagination
+            limit: How many items to return at most
+
+        Returns:
+            Graph result instance
+        """
+        query_builder = QueryBuilder.get()
+        query = query_builder.fetch_extracted_or_rule_items(
+            filter_by_query_string=bool(query_string),
+            filter_by_stable_target_id=bool(stable_target_id),
+        )
+        result = self.commit(
+            query,
+            query_string=query_string,
+            stable_target_id=stable_target_id,
+            labels=entity_type,
+            skip=skip,
+            limit=limit,
+        )
+        for query_result in result.all():
+            for item in query_result["items"]:
+                expand_references_in_search_result(item)
+        return result
+
     def fetch_extracted_items(
         self,
         query_string: str | None,
         stable_target_id: str | None,
-        entity_type: list[str] | None,
+        entity_type: Sequence[str] | None,
         skip: int,
         limit: int,
     ) -> Result:
@@ -175,24 +217,41 @@ class GraphConnector(BaseConnector):
         Returns:
             Graph result instance
         """
-        query_builder = QueryBuilder.get()
-        query = query_builder.fetch_extracted_items(
-            filter_by_query_string=bool(query_string),
-            filter_by_stable_target_id=bool(stable_target_id),
-            filter_by_labels=bool(entity_type),
+        return self._fetch_extracted_or_rule_items(
+            query_string,
+            stable_target_id,
+            entity_type or list(EXTRACTED_MODEL_CLASSES_BY_NAME),
+            skip,
+            limit,
         )
-        result = self.commit(
-            query,
-            query_string=query_string,
-            stable_target_id=stable_target_id,
-            labels=entity_type,
-            skip=skip,
-            limit=limit,
+
+    def fetch_rule_items(
+        self,
+        query_string: str | None,
+        stable_target_id: str | None,
+        entity_type: Sequence[str] | None,
+        skip: int,
+        limit: int,
+    ) -> Result:
+        """Query the graph for rule items.
+
+        Args:
+            query_string: Optional full text search query term
+            stable_target_id: Optional stable target ID filter
+            entity_type: Optional entity type filter
+            skip: How many items to skip for pagination
+            limit: How many items to return at most
+
+        Returns:
+            Graph result instance
+        """
+        return self._fetch_extracted_or_rule_items(
+            query_string,
+            stable_target_id,
+            entity_type or list(RULE_MODEL_CLASSES_BY_NAME),
+            skip,
+            limit,
         )
-        for query_result in result.all():
-            for extracted_item in query_result["items"]:
-                expand_references_in_search_result(extracted_item)
-        return result
 
     def fetch_identities(
         self,
@@ -356,8 +415,12 @@ class GraphConnector(BaseConnector):
             ref_positions=ref_positions,
         )
 
-    def create_rule(self, model: AnyRuleModel) -> AnyRuleModel:
-        """Create a new rule to be a applied to one merged item.
+    def create_rule_set(
+        self,
+        model: AnyRuleSetRequest | AnyRuleSetResponse,
+        stable_target_id: Identifier,
+    ) -> None:
+        """Create a new rule set to be applied to one merged item.
 
         This is a two-step process: first the rule and merged items are created
         along with their nested objects (like Text and Link); then all edges that
@@ -365,24 +428,16 @@ class GraphConnector(BaseConnector):
         the graph in a second step.
 
         Args:
-            model: A single rule model
-
-        Returns:
-            The created rule model
+            model: A rule-set model
+            stable_target_id: Identifier of the merged item
         """
-        stable_target_id = Identifier.generate()
-        self._merge_item(model, stable_target_id)
-        self._merge_edges(
-            model,
-            stable_target_id,
-            extra_refs=dict(
-                hadPrimarySource=MEX_PRIMARY_SOURCE_STABLE_TARGET_ID,
-                stableTargetId=stable_target_id,
-            ),
-        )
-        # TODO: read the rule back from the database instead of returning the
-        #       input; to ensure consistency (MX-1416)
-        return model
+        for rule in (model.additive, model.subtractive, model.preventive):
+            self._merge_item(rule, stable_target_id)
+            self._merge_edges(
+                rule,
+                stable_target_id,
+                extra_refs=dict(stableTargetId=stable_target_id),
+            )
 
     def ingest(self, models: list[AnyExtractedModel]) -> list[Identifier]:
         """Ingest a list of models into the graph as nodes and connect all edges.