+++ CHANGELOG.md
@@ -9,6 +9,9 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+- add `GraphConnector.fetch_rule_items` to get rule items
+- add `GET /rule-set/{stableTargetId}` endpoint to get rule-sets
+- add a rule-set response to the `create_rule_set` endpoint
 - implement merging logic as a triple of functions corresponding to our rule types
 - add a preview endpoint to perform merge with a submitted rule-set and all found items
 
@@ -23,6 +26,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - move searching and fetching of extracted and merged items to `helpers` module
   so they can be reused more easily outside of the endpoint code
 - use starlette status code constants instead of plain integers for readability
+- BREAKING: rework `create_rule` connector method and endpoint to use RuleSets
 
 ### Deprecated